+++ src/components/search/search.js
@@ -16,6 +16,7 @@ class Search extends React.Component {
           initialState: {
             resultsPerPage: 10
           },
+          trackUrlState: props.resultsOnly ? false : true, // do not push search query to URL if using resultsOnly
           searchQuery: {
             facets: {
               site: { type: 'value' }
@@ -52,7 +53,9 @@ class Search extends React.Component {
             reset
           }) => {
             return (
-              <div className="h36 relative">
+              <div
+                className={`${this.props.resultsOnly ? '' : 'h36 '}relative`}
+              >
                 <SearchBox
                   searchTerm={searchTerm}
                   trackClickThrough={trackClickThrough}
@@ -64,9 +67,14 @@ class Search extends React.Component {
                   inputId={props.inputId}
                   background={props.background}
                   narrow={props.narrow}
-                  disableModal={props.disableModal}
+                  disableModal={props.disableModal || props.resultsOnly} // disable modal if resultsOnly === true
                   site={props.site}
                   reset={reset}
+                  resultsOnly={props.resultsOnly}
+                  segmentTrackEvent={props.segmentTrackEvent}
+                  overrideSearchTerm={props.overrideSearchTerm}
+                  themeCompact={props.themeCompact}
+                  emptyResultMessage={props.emptyResultMessage}
                 />
               </div>
             );
@@ -78,13 +86,29 @@ class Search extends React.Component {
 }
 
 Search.propTypes = {
-  placeholder: PropTypes.string, // option to replace the input placehoder with a different string,
-  narrow: PropTypes.bool, // option to collapse input to fit in a crowded space
+  /** replace the input placehoder with a different string */
+  placeholder: PropTypes.string,
+  /** collapse input to fit in a crowded space */
+  narrow: PropTypes.bool, //
   background: PropTypes.oneOf(['light', 'dark']),
-  inputId: PropTypes.string, // option to override default id for input/label, used for testing
-  disableModal: PropTypes.bool, // option to completely disable modal if you always want an input
-  site: PropTypes.string, // option to add current site or all docs filter toggle
-  connector: PropTypes.instanceOf(SiteSearchAPIConnector) // option to connect to a custom search engine
+  /** override default id for input/label, used for testing */
+  inputId: PropTypes.string,
+  /** disable modal if you always want an input */
+  disableModal: PropTypes.bool,
+  /** add current site filter toggle */
+  site: PropTypes.string,
+  /** option to connect to a custom search engine */
+  connector: PropTypes.instanceOf(SiteSearchAPIConnector),
+  /** If true, only show results from search */
+  resultsOnly: PropTypes.bool,
+  /** If `resultsOnly: true` set `overrideSearchTerm` to set the search term */
+  overrideSearchTerm: PropTypes.string,
+  /** Segment track event, default is (Searched docs) */
+  segmentTrackEvent: PropTypes.string,
+  /** If true, enable compact mode utilizing smaller text and padding, default false */
+  themeCompact: PropTypes.bool,
+  /** Node to display when there are no search results for the given query */
+  emptyResultMessage: PropTypes.node
 };
 
 Search.defaultProps = {
@@ -95,7 +119,16 @@ Search.defaultProps = {
     engineKey: 'zpAwGSb8YMXtF9yDeS5K', // public engine key
     engineName: 'docs',
     documentType: ['page']
-  })
+  }),
+  resultsOnly: false,
+  segmentTrackEvent: 'Searched docs',
+  themeCompact: false,
+  emptyResultMessage: (
+    <p>
+      Hmmm, we didn't find anything. Reword your search, or{' '}
+      <a href="https://support.mapbox.com/hc/en-us">contact Support</a>.
+    </p>
+  )
 };
 
 export default Search;